Move SED stage into inference package
